Move all interation to 'bin/main.rb'
